backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
psycopg2-binary==2.9.9
django-environ==0.11.2
requests==2.31.0
newsdataapi==0.1.8
python-dateutil==2.8.2
langchain==0.1.0
langchain-google-genai==0.0.5
google-generativeai==0.3.2
numpy==1.24.3
scikit-learn==1.3.2
pgvector==0.2.4
django-cors-headers==4.3.1
django-filter==23.3
networkx==3.2.1
spacy==3.7.2
uk-core-news-sm @ https://github.com/explosion/spacy-models/releases/download/uk_core_news_sm-3.7.0/uk_core_news_sm-3.7.0-py3-none-any.whl